+++ src/Module/Monitor/BFER/Monitor_BFER.cpp
@@ -41,8 +41,8 @@ Monitor_BFER<B,R>
 	auto &ps_Y = this->template create_socket_in<R>(p2, "Y", this->N * this->n_frames);
 	this->create_codelet(p2, [this, &ps_X, &ps_Y]() -> int
 	{
-		return this->get_mutual_info(static_cast<B*>(ps_X.get_dataptr()),
-		                             static_cast<R*>(ps_Y.get_dataptr()));
+		return (int)this->get_mutual_info(static_cast<B*>(ps_X.get_dataptr()),
+		                                  static_cast<R*>(ps_Y.get_dataptr()));
 	});
 }
 
@@ -85,9 +85,9 @@ int Monitor_BFER<B,R>
 	int bit_errors_count;
 
 	if (count_unknown_values)
-		bit_errors_count = tools::hamming_distance_unk(U, V, this->K);
+		bit_errors_count = (int)tools::hamming_distance_unk(U, V, this->K);
 	else
-		bit_errors_count = tools::hamming_distance(U, V, this->K);
+		bit_errors_count = (int)tools::hamming_distance(U, V, this->K);
 
 	if (bit_errors_count)
 	{
